+++ tests/CoercerRegistryTest.php
@@ -39,6 +39,7 @@ class CoercerRegistryTest extends TestCase
 		$this->assertInstanceOf(Coercer::class, $integer);
 		$this->assertSame('Invalid number', $integer->message);
 		$this->assertInstanceOf(Coercer::class, $registry->get('float'));
+		$this->assertInstanceOf(Coercer::class, $registry->get('number'));
 		$this->assertInstanceOf(Coercer::class, $registry->get('list'));
 	}